backend/requirements.txt
fastapi
psycopg2-binary>=2.9.9
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9   
apscheduler>=3.10.4    
redis>=5.0.0
requests>=2.32.0
python-dotenv>=1.0.1
httpx==0.27.2
itsdangerous==2.2.0
uvicorn[standard]==0.38.0
websockets==15.0.1
wsproto>=1.2,<2
SQLAlchemy>=2.0
alembic>=1.13
